requirements.txt
awesome-slugify==1.6.5
Babel==2.3.4
beautifulsoup4==4.7.1
Flask==0.10.1
Flask-Babel==0.9
Flask-BasicAuth==0.2.0
Flask-Cache==0.13.1
Flask-Cors==2.1.2
Flask-Login==0.4.1
Flask-PyMongo==0.3.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
pymongo==2.9.5
pytz==2018.9
regex==2019.1.24
six==1.12.0
soupsieve==1.7.3
speaklater==1.3
Unidecode==1.0.23
Werkzeug==0.14.1
WTForms==2.2.1
